+++ meridian/analysis/analyzer.py
@@ -25,10 +25,10 @@ import warnings
 import arviz as az
 from meridian import backend
 from meridian import constants
+from meridian.common import errors
 from meridian.model import adstock_hill
 from meridian.model import context
 from meridian.model import equations
-from meridian.model import model
 from meridian.model import transformers
 import numpy as np
 import pandas as pd
@@ -59,7 +59,7 @@ def _validate_non_media_baseline_values_numbers(
 
 # TODO: Remove this method.
 def _get_model_context(
-    meridian: model.Meridian | None,
+    meridian: Any | None,
     model_context: context.ModelContext | None,
 ) -> context.ModelContext:
   """Gets `model_context`, handling the deprecated `meridian` argument."""
@@ -245,7 +245,8 @@ class DataTensors(backend.ExtensionType):
 
   def get_modified_times(
       self,
-      meridian: model.Meridian | None = None,
+      # TODO: Remove this argument.
+      meridian: Any | None = None,
       model_context: context.ModelContext | None = None,
   ) -> int | None:
     """Returns `n_times` of any tensor where `n_times` has been modified.
@@ -297,7 +298,8 @@ class DataTensors(backend.ExtensionType):
   def validate_and_fill_missing_data(
       self,
       required_tensors_names: Sequence[str],
-      meridian: model.Meridian | None = None,
+      # TODO: Remove this argument.
+      meridian: Any | None = None,
       model_context: context.ModelContext | None = None,
       allow_modified_times: bool = True,
   ) -> Self:
@@ -925,7 +927,7 @@ class Analyzer:
   def __init__(
       self,
       # TODO: Remove this argument.
-      meridian: model.Meridian | None = None,
+      meridian: Any | None = None,
       *,
       model_context: context.ModelContext | None = None,
       inference_data: az.InferenceData | None = None,
@@ -1664,7 +1666,7 @@ class Analyzer:
       )
     dist_type = constants.POSTERIOR if use_posterior else constants.PRIOR
     if dist_type not in self._inference_data.groups():
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           f"sample_{dist_type}() must be called prior to calling"
           " `expected_outcome()`."
       )
@@ -2162,7 +2164,7 @@ class Analyzer:
     dist_type = constants.POSTERIOR if use_posterior else constants.PRIOR
 
     if dist_type not in self.inference_data.groups():
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           f"sample_{dist_type}() must be called prior to calling this method."
       )
 
@@ -3820,7 +3822,7 @@ class Analyzer:
           "Must have at least one channel with reach and frequency data."
       )
     if dist_type not in self.inference_data.groups():
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           f"sample_{dist_type}() must be called prior to calling this method."
       )
 
@@ -4201,7 +4203,7 @@ class Analyzer:
         calling this method.
     """
     if constants.POSTERIOR not in self._inference_data.groups():
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           "sample_posterior() must be called prior to calling this method."
       )
 
@@ -4518,7 +4520,7 @@ class Analyzer:
         constants.PRIOR not in self._inference_data.groups()
         or constants.POSTERIOR not in self._inference_data.groups()
     ):
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           "sample_prior() and sample_posterior() must be called prior to"
           " calling this method."
       )
@@ -4983,7 +4985,7 @@ class Analyzer:
         constants.PRIOR not in self._inference_data.groups()
         or constants.POSTERIOR not in self._inference_data.groups()
     ):
-      raise model.NotFittedModelError(
+      raise errors.NotFittedModelError(
           "sample_prior() and sample_posterior() must be called prior to"
           " calling this method."
       )